+++ LargeFloat/test/multiplication.test.cpp
@@ -7,102 +7,102 @@
 #include <iostream>
 
 TEST(large_float, multiplication_zero) {
-    LargeNumbers::LargeFloatImplementation<uint8_t> number("0.0");
+    LargeNumbers::LargeFloatImplementation<uint32_t> number("0.0");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("1.0"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("1.0"));
     EXPECT_EQ(number.toString(), "0.0");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("-1.0"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("-1.0"));
     EXPECT_EQ(number.toString(), "0.0");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("0.1"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("0.1"));
     EXPECT_EQ(number.toString(), "0.0");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("-0.1"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("-0.1"));
     EXPECT_EQ(number.toString(), "0.0");
 
-    number = LargeNumbers::LargeFloatImplementation<uint8_t>("1.0");
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("0.0"));
+    number = LargeNumbers::LargeFloatImplementation<uint32_t>("1.0");
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("0.0"));
     EXPECT_EQ(number.toString(), "0.0");
 
-    number = LargeNumbers::LargeFloatImplementation<uint8_t>("-1.0");
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("0.0"));
+    number = LargeNumbers::LargeFloatImplementation<uint32_t>("-1.0");
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("0.0"));
     EXPECT_EQ(number.toString(), "0.0");
 }
 
 TEST(large_float, multiplication_without_fraction) {
-    LargeNumbers::LargeFloatImplementation<uint8_t> number("1.0");
+    LargeNumbers::LargeFloatImplementation<uint32_t> number("1.0");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("1.0"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("1.0"));
     EXPECT_EQ(number.toString(), "1.0");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("-1.0"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("-1.0"));
     EXPECT_EQ(number.toString(), "-1.0");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("-1.0"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("-1.0"));
     EXPECT_EQ(number.toString(), "1.0");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("2.0"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("2.0"));
     EXPECT_EQ(number.toString(), "2.0");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("-2.0"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("-2.0"));
     EXPECT_EQ(number.toString(), "-4.0");
 
-    number = LargeNumbers::LargeFloatImplementation<uint8_t>("2.0");
+    number = LargeNumbers::LargeFloatImplementation<uint32_t>("2.0");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("256.0"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("256.0"));
     EXPECT_EQ(number.toString(), "512.0");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("-2.0"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("-2.0"));
     EXPECT_EQ(number.toString(), "-1024.0");
 }
 
 TEST(large_float, multiplication_with_fraction) {
-    LargeNumbers::LargeFloatImplementation<uint8_t> number("0.1");
+    LargeNumbers::LargeFloatImplementation<uint32_t> number("0.1");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("0.1"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("0.1"));
     EXPECT_EQ(number.toString(), "0.01");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("0.01"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("0.01"));
     EXPECT_EQ(number.toString(), "0.0001");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("10000.0"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("10000.0"));
     EXPECT_EQ(number.toString(), "1.0");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("0.0625"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("0.0625"));
     EXPECT_EQ(number.toString(), "0.0625");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("0.0625"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("0.0625"));
     EXPECT_EQ(number.toString(), "0.00390625");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("0.1"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("0.1"));
     EXPECT_EQ(number.toString(), "0.000390625");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("10.0"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("10.0"));
     EXPECT_EQ(number.toString(), "0.00390625");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("-2.0"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("-2.0"));
     EXPECT_EQ(number.toString(), "-0.0078125");
 }
 
 TEST(large_float, multiplication_mixed) {
-    LargeNumbers::LargeFloatImplementation<uint8_t> number("1.1");
+    LargeNumbers::LargeFloatImplementation<uint32_t> number("1.1");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("1.1"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("1.1"));
     EXPECT_EQ(number.toString(), "1.21");
 
-    number.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("-2.223"));
+    number.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("-2.223"));
     EXPECT_EQ(number.toString(), "-2.68983");
 }
 
-LargeNumbers::LargeFloatImplementation<uint8_t> createNumber(const std::string &number) {
-    LargeNumbers::LargeFloatImplementation<uint8_t> n(number);
+LargeNumbers::LargeFloatImplementation<uint32_t> createNumber(const std::string &number) {
+    LargeNumbers::LargeFloatImplementation<uint32_t> n(number);
     EXPECT_EQ(n.toString(), number);
     return n;
 }
 
 TEST(large_float, high_precision_multiplication_1) {
-    LargeNumbers::LargeFloatImplementation<uint8_t>::setDecimalPrecision(100);
+    LargeNumbers::LargeFloatImplementation<uint32_t>::setDecimalPrecision(100);
 
     auto n1 = createNumber(
             "0.1111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111");
@@ -134,11 +134,11 @@ TEST(large_float, high_precision_multiplication_2) {
 }
 
 TEST(large_float, high_precision_multiplication_3) {
-    LargeNumbers::LargeFloatImplementation<uint8_t>::setDecimalPrecision(100);
+    LargeNumbers::LargeFloatImplementation<uint32_t>::setDecimalPrecision(100);
 
-    LargeNumbers::LargeFloatImplementation<uint8_t> n1(
+    LargeNumbers::LargeFloatImplementation<uint32_t> n1(
             "2.3333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333");
-    LargeNumbers::LargeFloatImplementation<uint8_t> n2(
+    LargeNumbers::LargeFloatImplementation<uint32_t> n2(
             "3.1111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111");
 
     n1.multiply(n2);
@@ -154,6 +154,6 @@ TEST(large_float, multiplication_large_numbers) {
     n1.multiply(n2);
     EXPECT_EQ(n1.toString(), "3594675427188164949046079279342197.315943069945180624309620243372");
 
-    n1.multiply(LargeNumbers::LargeFloatImplementation<uint8_t>("0.0"));
+    n1.multiply(LargeNumbers::LargeFloatImplementation<uint32_t>("0.0"));
     EXPECT_EQ(n1.toString(), "0.0");
 }
